% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_prices}
\alias{summarize_prices}
\title{Summarizes price column from provided data frame / tibble}
\usage{
summarize_prices(tbl)
}
\arguments{
\item{tbl}{data to be summarized}
}
\value{
sum of the price column
}
\description{
Summarizes price column from provided data frame / tibble
}
